skills/pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[project]
name = "rusticai-skills"
version = "0.1.0"
description = "Agent Skills support for Rustic AI - load, parse, and execute skills from the Agent Skills ecosystem"
authors = [{name = "Dragonscale Industries Inc.", email = "dev@dragonscale.ai"}]
license = "Apache-2.0"
readme = "README.md"
requires-python = ">=3.12,<3.13"
urls = { Homepage = "https://www.rustic.ai/", Repository = "https://github.com/dragonscale-ai/rustic-ai", "Rustic AI Core" = "https://pypi.org/project/rusticai-core/" }


[tool.poetry]
packages = [{ include = "rustic_ai", from = "src" }]

[tool.poetry.dependencies]
rusticai-core = { path = "../core", develop = true }
rusticai-llm-agent = { path = "../llm-agent", develop = true }
gitpython = "^3.1.0"
pyyaml = "^6.0.0"
httpx = "^0.28.0"

[tool.poetry-monorepo.deps]


[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
black = { version = "^24.2.0", extras = ["jupyter"] }
flake8 = "^7.1.2"
tox = "^4.24.1"
isort = "^6.0.0"
mypy = "^1.15.0"
flaky = "^3.8.1"
pytest-asyncio = "^0.26.0"
rusticai-testing = {path = "../testing", develop = true}

[tool.poetry.extras]
test = ["pytest"]

[tool.black]
line-length = 120
target-version = ['py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
check_untyped_defs = true
plugins = "pydantic.mypy"
explicit_package_bases = true


[tool.isort]
profile = "black"
force_sort_within_sections = true # For the sorting order issues
lines_between_sections = 1 # For the newline issue
known_third_party = ["pydantic", "git", "yaml", "httpx"]
known_first_party = ["rustic_ai.core", "rustic_ai.llm_agent"]

